requirements.txt
playwright==1.43.0
neo4j==5.19.0
python-dotenv==1.0.1
openai>=1.0.0
